Added some code for the Cell-Implementation of the reduction to sum operation, plus some helping structures.
git-svn-id: file:///home/pg512/share/svn/pg512/trunk@1483 0f804a50-a62f-0410-84e1-872c020a11f9

+++ honei/cell/libutil/operations.hh
@@ -202,7 +202,7 @@ namespace honei
          * \ingroup grpframework
          */
         float sum_float(const vector float & value);
-        double sum_double(const vector float & value);
+        double sum_double(const vector double & value);
 
         /**
          * Return the maximum of all of the vector's elements.
